# specs/001-warehouse-management/contracts/openapi.yaml
openapi: 3.1.2
...
info:
  title: Warehouse Manager API
  version: 1.0.0
  description: >-
    Authoritative HTTP contract for warehouse, route, sale, reconciliation, reporting, document, user, and
    printer-profile operations. Money and quantities are decimal strings. All protected authorization and business rules
    are enforced by the API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /routes:
    get:
      tags: [ Routes ]
      operationId: listRoutes
      description: >-
        Administrators may list all routes. Drivers receive every route assigned to them, including Returned and Closed
        history. A Driver-provided filter cannot expand visibility to another Driver's routes.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/RouteState"
        - name: driverId
          in: query
          description: Administrator filter; Drivers may specify only their own identity.
          schema:
            $ref: "#/components/schemas/Uuid"
        - name: businessDate
          in: query
          schema:
            type: string
            format: date
        - name: active
          in: query
          description: True returns non-closed routes; false returns closed history. Omitted returns both, within the
            authenticated role scope.
          schema:
            type: boolean
      responses:
        '200':
          description: Routes visible under the caller's role and resource scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteListResponse'
        '403':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List routes
    post:
      tags: [ Routes ]
      operationId: createRoute
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreateRequest'
      responses:
        '201':
          description: Preparing route created and assigned by an administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create route
  /routes/{routeId}:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      tags: [ Routes ]
      operationId: getRoute
      description: >-
        Administrators may retrieve any route. Drivers may retrieve only routes assigned to them, in any state; another
        Driver's route is forbidden.
      responses:
        '200':
          description: Route with load, balances, movements, sales, return, reconciliation, and closure history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetailResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Get route
  /routes/{routeId}/load:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    put:
      tags: [ Routes ]
      operationId: replaceRouteLoadDraft
      description: Assigned driver records the complete draft load while route is Preparing.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteLoadDraftRequest'
      responses:
        '200':
          description: Draft load replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteLoadResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Replace route load draft
  /routes/{routeId}/load/confirmation:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [ Routes ]
      operationId: confirmRouteLoad
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpectedVersionRequest'
      responses:
        '200':
          description: Full load moved from branch to temporary route inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetailResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Confirm route load
  /routes/{routeId}/start:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [ Routes ]
      operationId: startRoute
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpectedVersionRequest'
      responses:
        '200':
          description: Assigned driver moved route from Preparing to En Route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Start route
  /routes/{routeId}/return:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [ Routes ]
      operationId: returnRoute
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteReturnRequest"
      responses:
        '200':
          description: Assigned driver moved route from En Route to Returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Return route
  /routes/{routeId}/reconciliation:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    put:
      tags: [ Routes ]
      operationId: approveRouteReconciliation
      description: >-
        Administrator records physical returns, provides mandatory reasons for differences, creates adjustments, and
        returns stock to the origin atomically.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteReconciliationRequest'
      responses:
        '200':
          description: Reconciliation approved and temporary route inventory is zero
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteReconciliationResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Approve route reconciliation
  /routes/{routeId}/close:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    post:
      tags: [ Routes ]
      operationId: closeRoute
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpectedVersionRequest'
      responses:
        '200':
          description: Fully reconciled route closed by an administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Close route
...
  /routes/{routeId}/return-declaration:
    parameters:
      - $ref: "#/components/parameters/RouteId"
    post:
      tags:
        - Routes
      operationId: declareRouteReturn
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteReconciliationRequest"
      responses:
        "200":
          description: Return declaration saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Declare physical quantities for a returned route
      description: Assigned driver only. Records an immutable declaration before administrator reconciliation, without
        changing stock. Supports routes returned before quantity entry was available.
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-wm_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Per-session synchronizer token for an authenticated unsafe request
      schema:
        type: string
        minLength: 32
        maxLength: 512
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key scoped to actor and operation; reuse requires identical content
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    RouteId:
      name: routeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    Problem:
      description: RFC 9457 error; exact status and stable code identify the failure
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Uuid:
      type: string
      format: uuid
...
    ExpectedVersionRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
    RouteState:
      type: string
      enum: [ PREPARING, EN_ROUTE, RETURNED, CLOSED ]
    RouteCreateRequest:
      type: object
      additionalProperties: false
      required: [ originLocationId, driverId, vehicleId, businessDate ]
      properties:
        originLocationId:
          $ref: '#/components/schemas/Uuid'
        driverId:
          $ref: '#/components/schemas/Uuid'
        vehicleId:
          $ref: '#/components/schemas/Uuid'
        businessDate:
          type: string
          format: date
        routeNumber:
          type: string
          minLength: 1
          maxLength: 64
...
    RouteResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Route'
    RouteListResponse:
      type: object
      required: [ data, page ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Route'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    RouteLoadDraftRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion, lines ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        lines:
          type: array
          minItems: 1
          maxItems: 200
          items:
            $ref: '#/components/schemas/RouteLoadLineRequest'
...
    RouteLoadResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/RouteLoad'
...
    RouteDetailResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/RouteDetail'
...
    RouteReturnRequest:
      type: object
      properties:
        expectedVersion:
          type: integer
          exclusiveMinimum: 0
          maximum: 9007199254740991
        lines:
          minItems: 1
          maxItems: 200
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
                format: uuid
                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              physicalReturnQuantity:
                type: string
                pattern: ^(0|[1-9]\d*)(\.\d{1,3})?$
              differenceReason:
                anyOf:
                  - type: string
                    minLength: 1
                    maxLength: 500
                  - type: "null"
            required:
              - productId
              - physicalReturnQuantity
            additionalProperties: false
      required:
        - expectedVersion
      additionalProperties: false
...